[CREATE] TeamCalendarViewController file
